refactor: extract shared panel toggle hooks to reduce duplication
Toggle logic for sidebar, terminal, and chat panels was duplicated
across status-bar, watermark, chat-container, and keyboard shortcuts.
