+++ web/components/project/chat/components/chat-container.tsx
@@ -6,7 +6,7 @@ import {
   TooltipContent,
   TooltipTrigger,
 } from "@/components/ui/tooltip"
-import { useEditor } from "@/context/editor-context"
+import { useToggleChat } from "@/components/project/layout/hooks/usePanelToggles"
 import { cn } from "@/lib/utils"
 import { Slot } from "@radix-ui/react-slot"
 import { type VariantProps } from "class-variance-authority"
@@ -349,11 +349,7 @@ function ScrollButton({
 }
 
 export function ChatContainerCollapse() {
-  const { gridRef } = useEditor()
-  function toggleAIChat() {
-    const panel = gridRef.current?.getPanel("chat")
-    panel?.api.setVisible(!panel.api.isVisible)
-  }
+  const toggleAIChat = useToggleChat()
   const { maximized } = useChatContainerContext()
 
   return (
